src/chatbot/scraper.py
import requests
from bs4 import BeautifulSoup

class SahibindenScraper:
    def __init__(self, base_url="https://www.sahibinden.com"):
        self.base_url = base_url

    def search(self, query):
        url = f"{self.base_url}/{query}"
        response = requests.get(url)
        soup = BeautifulSoup(response.text, 'html.parser')
        # Extract the data from the soup object
        # This will depend on the structure of the Sahibinden.com website
        data = self._extract_data(soup)
        return data

    def _extract_data(self, soup):
        # This method should be implemented based on the structure of the Sahibinden.com website
        pass
